+++ lib/type_result.dart
@@ -2,3 +2,4 @@ library type_result;
 
 export 'src/result.dart';
 export 'src/future_result.dart';
+export 'src/generics_extensions.dart';