+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 ## 1.0.2
 
 * Fix import
+
+## 1.0.3
+
+* Add some method to FutureResult 
+
+## 1.0.4
+
+* Add some method to FutureResult